+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/representations/PlayerRepresentation.java
@@ -18,11 +18,21 @@ public class PlayerRepresentation {
 
   private String name;
 
+  private String displayName;
+
   private String domain;
 
   private String duplicatePlayerName;
 
   private boolean bot;
+
+  public String getDisplayName() {
+    return displayName;
+  }
+
+  public void setDisplayName(String displayName) {
+    this.displayName = displayName;
+  }
 
   public boolean isBot() {
     return bot;