+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 ## Release Notes
 
+
+**Bot Preferences**
+
+- Added option to disable bot commands.
+- Added more user friendly for user allow list creation.
+
+
 **Bug Fixes**
 
 - Fixed errorneous time field initialization for existing competitions.